Just use the code from dcicutils. Incorporate transitional tests in single test file.
